moved health, cleanup and routes to elife-app component
